Move data handling functions in models
