+++ envs/secrets.env.template
@@ -28,6 +28,6 @@ MAILJET_WEBHOOK_SECRET="16 random char:16 other random chars"
 APM_SERVER_URL=set_it
 APM_AUTH_TOKEN=set_it
 
-# Ddefault values of the integration platform are not confidentials
+# Default values of the integration platform are not confidentials
 FRANCE_CONNECT_CLIENT_ID=set_it
 FRANCE_CONNECT_CLIENT_SECRET=set_it
